+++ UIKit/SpecHelper/Extensions/UICollectionViewCell+Spec.m
@@ -1,4 +1,6 @@
 #import "UICollectionViewCell+Spec.h"
+#import "UIView+Spec.h"
+#import "../Helpers/PCKPrototypeCellInstantiatingDataSource.h"
 
 @interface UICollectionView (PrivateAppleMethods)
 - (void)_userSelectItemAtIndexPath:(NSIndexPath *)indexPath;
@@ -17,6 +19,25 @@
 
     NSIndexPath *indexPath = [collectionView indexPathForCell:self];
     [collectionView _userSelectItemAtIndexPath:indexPath];
+}
+
++ (instancetype)instantiatePrototypeCellFromStoryboard:(UIStoryboard *)storyboard
+                              viewControllerIdentifier:(NSString *)viewControllerIdentifier
+                                 collectionViewKeyPath:(NSString *)collectionViewKeyPath
+                                        cellIdentifier:(NSString *)cellIdentifier {
+    NSAssert(storyboard, @"Must provide a storyboard");
+    NSAssert([cellIdentifier length] > 0, @"Must provide a cell identifier");
+
+    UIViewController *viewController = viewControllerIdentifier ? [storyboard instantiateViewControllerWithIdentifier:viewControllerIdentifier] : [storyboard instantiateInitialViewController];
+    NSAssert(viewController, @"Could not find the view controller");
+
+    [viewController view];
+
+    UICollectionView *collectionView = collectionViewKeyPath ? [viewController valueForKeyPath:collectionViewKeyPath] : [viewController.view firstSubviewOfClass:[UICollectionView class]];
+    NSAssert(collectionView, @"Could not find the collection view");
+
+    PCKPrototypeCellInstantiatingDataSource *dataSource = [[[PCKPrototypeCellInstantiatingDataSource alloc] initWithCollectionView:collectionView] autorelease];
+    return [dataSource collectionViewCellWithIdentifier:cellIdentifier];
 }
 
 @end